fix: fixes zip closed error
