+++ .changeset/four-planes-type.md
@@ -1,0 +1,5 @@
+---
+'@yam-js/test-util': patch
+---
+
+adds getLogs method to server and cleanups code
