+++ .changeset/seven-badgers-kick.md
@@ -1,0 +1,5 @@
+---
+'@yam-js/plugin': patch
+---
+
+fixes the zip file closed error
